just noticed last commit didnt work correctly

+++ injector/parsing.h
@@ -14,7 +14,7 @@ struct IMAGE_DATA
 struct MODULE
 {
 	HMODULE handle = nullptr;
-	DWORD ImageBase = NULL;
+	DWORD64 ImageBase = NULL;
 	IMAGE_DATA image;
 };
 
@@ -52,7 +52,7 @@ int ResolveImports(const IMAGE_DATA* const target);
 #define METHOD_HIJACK_THREAD 0x10
 
 
-template <typename ret> auto ConvertRva(const void* const base, const DWORD rva, const IMAGE_DATA* const image)->ret
+template <typename ret> auto ConvertRva(const void* const base, const DWORD64 rva, const IMAGE_DATA* const image)->ret
 {
 	const IMAGE_SECTION_HEADER* SectionHeader = image->sections;
